app/requirements.txt
asgiref==3.3.4
Django>=3.0,<4.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
gunicorn==20.0.4
psycopg2==2.8.6